fixed convert_to_json

+++ players/utils.py
@@ -1,12 +1,19 @@
 import json
+import os.path
+
 import pandas as pd
 
 
 def convert_to_json():
-    all_player_stats_df = pd.read_csv('data/all_player_stats.csv')
-    all_player_stats_df.to_json('all_player_stats.json', orient='records', lines=False)
-    league_averages_df = pd.read_csv('data/league_averages.csv')
-    league_averages_df.to_json('league_averages.json')
+
+    if not os.path.exists('all_player_stats.json'):
+        all_player_stats_df = pd.read_csv('data/all_player_stats.csv')
+        all_player_stats_df.to_json('all_player_stats.json', orient='records', lines=False)
+
+
+    if not os.path.exists('league_averages.json'):
+        league_averages_df = pd.read_csv('data/league_averages.csv')
+        league_averages_df.to_json('league_averages.json', orient='records', lines=False)
 
 
 def load_player_data():
